+++ components/TooltipButton.vue
@@ -1,7 +1,7 @@
 <template>
   <v-tooltip bottom>
     <template #activator="{ on: onTooltip, attrs }">
-      <v-btn :to="to ? to : ''" icon @click="onClicked">
+      <v-btn :to="to ? to : ''" icon v-bind="$attrs" v-on="$listeners">
         <v-icon v-bind="attrs" class="text--primary" v-on="onTooltip">
           {{ icon }}
         </v-icon>
@@ -14,17 +14,8 @@
 export default {
   props: {
     icon: String,
-    clicked: Function,
     hover: String,
     to: String,
-    clickarg: String,
-  },
-  methods: {
-    onClicked() {
-      if (this.clicked !== null)
-        if (this.clickarg !== null) this.clicked(this.clickarg)
-        else this.clicked()
-    },
   },
 }
 </script>
