+++ store/index.js
@@ -4,9 +4,16 @@ export const state = () => ({
   username: '',
   token: '',
   dark: true,
-  loggedIn: false
+  loggedIn: false,
 })
 
 export const mutations = {
-  ...createMutations('username', 'token', 'dark', 'loggedIn')
+  ...createMutations('username', 'token', 'dark', 'loggedIn'),
+}
+
+export const actions = {
+  nuxtServerInit(_, { $auth, $vuetify }) {
+    // updating vuetify theme here so the theme is consistend during SSR
+    $vuetify.theme.dark = $auth.$storage.getUniversal('isDarkTheme')
+  },
 }
